Add files via upload
This project is made for my college Thapar University. If a new person enters college, he doesn't know the path.  This helps user find the path. It even tells the places where a person can eat if hungry or relax or any othe query. It even answers the queries of the person regarding college.